# Dynamic send and batch send splits for milestone 1.5