--- a/pkg/message/interface.go
+++ b/pkg/message/interface.go
@@ -2,6 +2,10 @@ package message
 
 import "github.com/rabbitmq/rabbitmq-stream-go-client/pkg/amqp"
 
+// StreamMessage is the interface that wraps the basic methods to interact with a message
+// in the context of a stream.
+// Currently, the StreamMessage interface is implemented by the amqp.Message struct.
+// The implementations are not meant to be thread-safe.
 type StreamMessage interface {
 	MarshalBinary() ([]byte, error)
 	UnmarshalBinary(data []byte) error