--- a/pkg/stream/blocking_queue.go
+++ b/pkg/stream/blocking_queue.go
@@ -2,6 +2,7 @@ package stream
 
 import (
 	"errors"
+	"github.com/rabbitmq/rabbitmq-stream-go-client/pkg/logs"
 	"sync/atomic"
 	"time"
 )
@@ -9,17 +10,15 @@ import (
 var ErrBlockingQueueStopped = errors.New("blocking queue stopped")
 
 type BlockingQueue[T any] struct {
-	queue    chan T
-	capacity int
-	status   int32
+	queue  chan T
+	status int32
 }
 
 // NewBlockingQueue initializes a new BlockingQueue with the given capacity
 func NewBlockingQueue[T any](capacity int) *BlockingQueue[T] {
 	return &BlockingQueue[T]{
-		queue:    make(chan T, capacity),
-		capacity: capacity,
-		status:   0,
+		queue:  make(chan T, capacity),
+		status: 0,
 	}
 }
 
@@ -28,7 +27,8 @@ func (bq *BlockingQueue[T]) Enqueue(item T) error {
 	if bq.IsStopped() {
 		return ErrBlockingQueueStopped
 	}
-	bq.queue <- item // This will block if the queue is full
+	bq.queue <- item
+
 	return nil
 }
 
@@ -39,7 +39,11 @@ func (bq *BlockingQueue[T]) Dequeue(timeout time.Duration) T {
 		return zeroValue
 	}
 	select {
-	case item := <-bq.queue:
+	case item, ok := <-bq.queue:
+		if !ok {
+			var zeroValue T // Zero value of type T
+			return zeroValue
+		}
 		return item
 	case <-time.After(timeout):
 		var zeroValue T // Zero value of type T
@@ -56,12 +60,30 @@ func (bq *BlockingQueue[T]) IsEmpty() bool {
 }
 
 // Stop stops the queue from accepting new items
-// but allows the existing items to be processed
+// but allows some pending items.
 // Stop is different from Close in that it allows the
 // existing items to be processed.
-// That avoids the need to drain the queue before closing it.
+// Drain the queue to be sure there are not pending messages
 func (bq *BlockingQueue[T]) Stop() {
 	atomic.StoreInt32(&bq.status, 1)
+	// drain the queue. To be sure there are not pending messages
+	// in the queue.
+	// it does not matter if we lose some messages here
+	// since there is the unConfirmed map to handle the messages
+	isActive := true
+	for isActive {
+		select {
+		case <-bq.queue:
+			// do nothing
+		case <-time.After(10 * time.Millisecond):
+			isActive = false
+			return
+		default:
+			isActive = false
+			return
+		}
+	}
+	logs.LogDebug("BlockingQueue stopped")
 }
 
 func (bq *BlockingQueue[T]) Close() {