--- a/pkg/stream/client.go
+++ b/pkg/stream/client.go
@@ -756,8 +756,8 @@ func (c *Client) BrokerForConsumer(stream string) (*Broker, error) {
 		brokers = append(brokers, replica)
 	}
 
-	rand.Seed(time.Now().UnixNano())
-	n := rand.Intn(len(brokers))
+	r := rand.New(rand.NewSource(SEED))
+	n := r.Intn(len(brokers))
 	return brokers[n], nil
 }