1.5 Milestone
----
The PR contains code refactoring. Focus on increasing the stability during the reconnection and introducing the dynamic send.


News:
- Update to Golang 1.22.0

- Removed the aggregation via timeout and implemented the dynamic send. The goal is to reduce the latency, especially in low-rate use cases. Low-rate use cases are the most used. 
The dynamic send is a wrapper around the `BatchSend()` that will remain public for the users who want more control during the send. 

- Refactor Producer close. Remove the code duplication when the producer is closed correctly or not
 
- Make the unconfirmed operation more atomic to avoid rare cases when the unconfirmed operations were out of the mutex lock

- Increase the timeout for the Reliable producers and consumers. It was too low. Now, there is one more wait for the producer to wait for pending messages to be flushed when closed. 

- Refactor publish confirm channel with a lock to check if it is valid or not and avoid race conditions during the closing

- Handle when the producer is not found during the confirmation. It can happen when the producer is closed before all the messages are confirmed

Test results 
-----
PerTest with a limit rate of 50 Messages per second

Given:
```
 ./perftest --streams stream --publishers 1 --consumers 1 --max-length-bytes 10GB --async-send --rate 50;
```


With this PR the latency is around 5 ms
```
Published     40.0 msg/s | Confirmed     40.0 msg/s |  Consumed     40.0 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 5 ms
Published     41.6 msg/s | Confirmed     41.6 msg/s |  Consumed     41.6 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 4 ms
Published     42.9 msg/s | Confirmed     42.9 msg/s |  Consumed     42.9 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 4 ms
Published     43.8 msg/s | Confirmed     43.8 msg/s |  Consumed     43.8 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 4 ms
```

Main branch (`1.4.x`), the latency is around 90 ms ( The timeout for the aggregation ) 
```
Published     44.4 msg/s | Confirmed     44.4 msg/s |  Consumed     44.4 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 108 ms
Published     45.0 msg/s | Confirmed     45.0 msg/s |  Consumed     45.0 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 106 ms
Published     45.5 msg/s | Confirmed     45.5 msg/s |  Consumed     45.5 msg/s |  Rate Fx: 50 | Body sz: 8 | latency: 103 ms
```


With the dynamic send, the latency is reduced, especially with a low rate, which is the most common use case.